(NFC) Move passes to their own dialect
