+++ src/Dialect/ONNX/Transforms/InstrumentONNXSignaturePass.cpp
@@ -2,15 +2,14 @@
  * SPDX-License-Identifier: Apache-2.0
  */
 
-//===------- InstrumentONNXSignaturePass.cpp - Instrumentation
-//---------------------===//
+//===------- InstrumentONNXSignaturePass.cpp - Instrumentation ------------===//
 //
 // Copyright 2022 The IBM Research Authors.
 //
 // =============================================================================
 //
-// This file implements a Function level pass that inserts krnl print statements
-// that print the operation name and its input type signature at runtime.
+// This file implements a Function level pass that inserts statements that print
+// the operation name and its input type signature at runtime.
 //
 //===----------------------------------------------------------------------===//
 
@@ -25,8 +24,6 @@
 #include "llvm/ADT/SmallPtrSet.h"
 #include "llvm/Support/raw_ostream.h"
 
-#include "src/Dialect/Krnl/DialectBuilder.hpp"
-#include "src/Dialect/Krnl/KrnlOps.hpp"
 #include "src/Dialect/ONNX/ONNXOps.hpp"
 #include "src/Dialect/ONNX/ONNXOps/OpHelper.hpp"
 #include "src/Interface/ShapeInferenceOpInterface.hpp"
@@ -37,7 +34,7 @@ using namespace mlir;
 namespace {
 
 /*!
- * This pass insert KrnlPrint and KrnlPrintTensor before each ONNX ops to print
+ * This pass insert ONNXPrintSignatureOp before each ONNX ops to print
  * an operation name and input operand type signatures at runtime.
  */
